Refactor file upload jobs
