+++ storage/inkless/src/main/java/io/aiven/inkless/storage_backend/common/ObjectUploader.java
@@ -17,10 +17,11 @@
  */
 package io.aiven.inkless.storage_backend.common;
 
+import java.io.InputStream;
+
 import io.aiven.inkless.common.ObjectKey;
-import io.aiven.inkless.merge.MergeBatchInputStream;
 
 public interface ObjectUploader {
     void upload(ObjectKey key, byte[] data) throws StorageBackendException;
-    void upload(ObjectKey key, MergeBatchInputStream data) throws StorageBackendException;
+    void uploadMultiPart(ObjectKey key, InputStream data) throws StorageBackendException;
 }
